Split tem tests and update
